+++ server/match_registry.go
@@ -41,7 +41,8 @@ var (
 
 	MaxLabelSize = 2048
 
-	ErrMatchLabelTooLong = errors.New("match label too long, must be 0-2048 bytes")
+	ErrMatchLabelTooLong     = errors.New("match label too long, must be 0-2048 bytes")
+	ErrDeferredBroadcastFull = errors.New("too many deferred message broadcasts per tick")
 )
 
 type MatchIndexEntry struct {
@@ -89,7 +90,7 @@ type MatchRegistry interface {
 	// Create and start a new match, given a Lua module name or registered Go match function.
 	CreateMatch(ctx context.Context, logger *zap.Logger, createFn RuntimeMatchCreateFunction, module string, params map[string]interface{}) (string, error)
 	// Register and initialise a match that's ready to run.
-	NewMatch(logger *zap.Logger, id uuid.UUID, label *atomic.String, core RuntimeMatchCore, params map[string]interface{}) (*MatchHandler, error)
+	NewMatch(logger *zap.Logger, id uuid.UUID, core RuntimeMatchCore, params map[string]interface{}) (*MatchHandler, error)
 	// Return a match handler by ID, only from the local node.
 	GetMatch(id uuid.UUID) *MatchHandler
 	// Remove a tracked match and ensure all its presences are cleaned up.
@@ -158,16 +159,8 @@ func NewLocalMatchRegistry(logger, startupLogger *zap.Logger, config Config, tra
 func (r *LocalMatchRegistry) CreateMatch(ctx context.Context, logger *zap.Logger, createFn RuntimeMatchCreateFunction, module string, params map[string]interface{}) (string, error) {
 	id := uuid.Must(uuid.NewV4())
 	matchLogger := logger.With(zap.String("mid", id.String()))
-	label := atomic.NewString("")
-	labelUpdateFn := func(input string) error {
-		if err := r.UpdateMatchLabel(id, input); err != nil {
-			return err
-		}
-		label.Store(input)
-		return nil
-	}
 
-	core, err := createFn(ctx, matchLogger, id, r.node, module, labelUpdateFn)
+	core, err := createFn(ctx, matchLogger, id, r.node, module)
 	if err != nil {
 		return "", err
 	}
@@ -176,7 +169,7 @@ func (r *LocalMatchRegistry) CreateMatch(ctx context.Context, logger *zap.Logger
 	}
 
 	// Start the match.
-	mh, err := r.NewMatch(matchLogger, id, label, core, params)
+	mh, err := r.NewMatch(matchLogger, id, core, params)
 	if err != nil {
 		return "", fmt.Errorf("error creating match: %v", err.Error())
 	}
@@ -184,13 +177,13 @@ func (r *LocalMatchRegistry) CreateMatch(ctx context.Context, logger *zap.Logger
 	return mh.IDStr, nil
 }
 
-func (r *LocalMatchRegistry) NewMatch(logger *zap.Logger, id uuid.UUID, label *atomic.String, core RuntimeMatchCore, params map[string]interface{}) (*MatchHandler, error) {
+func (r *LocalMatchRegistry) NewMatch(logger *zap.Logger, id uuid.UUID, core RuntimeMatchCore, params map[string]interface{}) (*MatchHandler, error) {
 	if r.stopped.Load() {
 		// Server is shutting down, reject new matches.
 		return nil, errors.New("shutdown in progress")
 	}
 
-	match, err := NewMatchHandler(logger, r.config, r, core, label, id, r.node, params)
+	match, err := NewMatchHandler(logger, r.config, r, core, id, r.node, params)
 	if err != nil {
 		return nil, err
 	}