+++ server/runtime_go.go
@@ -1743,7 +1743,7 @@ func NewRuntimeProviderGo(logger, startupLogger *zap.Logger, db *sql.DB, config 
 
 	match := make(map[string]func(ctx context.Context, logger runtime.Logger, db *sql.DB, nk runtime.NakamaModule) (runtime.Match, error), 0)
 	matchLock := &sync.RWMutex{}
-	matchCreateFn := func(ctx context.Context, logger *zap.Logger, id uuid.UUID, node string, name string, labelUpdateFn RuntimeMatchLabelUpdateFunction) (RuntimeMatchCore, error) {
+	matchCreateFn := func(ctx context.Context, logger *zap.Logger, id uuid.UUID, node string, name string) (RuntimeMatchCore, error) {
 		matchLock.RLock()
 		fn, ok := match[name]
 		matchLock.RUnlock()
@@ -1758,8 +1758,7 @@ func NewRuntimeProviderGo(logger, startupLogger *zap.Logger, db *sql.DB, config 
 			return nil, err
 		}
 
-		//return NewRuntimeGoMatchCore(logger, matchRegistry, tracker, router, id, node, labelUpdateFn, db, env, nk, match)
-		return NewRuntimeGoMatchCore(logger, matchRegistry, router, id, node, labelUpdateFn, db, env, nk, match)
+		return NewRuntimeGoMatchCore(logger, matchRegistry, router, id, node, db, env, nk, match)
 	}
 	nk.SetMatchCreateFn(matchCreateFn)
 	matchNamesListFn := func() []string {